2020-cleanup of output and fixed broken test/day

+++ src/test/java/year2020/day22/Day22Test.java
@@ -28,7 +28,7 @@ class Day22Test {
                 """;
         var values = Arrays.stream(input.split("\n")).collect(Collectors.toList());
 
-        assertEquals(306, Day22.solveA(values));
+        assertEquals(306L, (Long) Day22.solveA(values));
     }
 
     @Test
